Fix GetEnterpriseValueAsync model (#37)
* fix type

* Bump version

+++ FinancialModelingPrepApi/Model/CompanyValuation/EnterpriseValueResponse.cs
@@ -20,10 +20,10 @@ namespace MatthiWare.FinancialModelingPrep.Model.CompanyValuation
         public double MarketCapitalization { get; set; }
 
         [JsonPropertyName("minusCashAndCashEquivalents")]
-        public long MinusCashAndCashEquivalents { get; set; }
+        public double MinusCashAndCashEquivalents { get; set; }
 
         [JsonPropertyName("addTotalDebt")]
-        public long AddTotalDebt { get; set; }
+        public double AddTotalDebt { get; set; }
 
         [JsonPropertyName("enterpriseValue")]
         public double EnterpriseValue { get; set; }
